requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
python-multipart
tensorflow>=2.13
scikit-learn
matplotlib
icrawler
Pillow